requirements.txt
instructure-dap-client >=2.0.0
python-dotenv
psycopg2-binary
sqlalchemy
pysqlsync[postgresql]